+++ polar-core/src/events.rs
@@ -1,0 +1,66 @@
+use super::kb::*;
+use super::terms::*;
+use super::traces::*;
+use serde::{Deserialize, Serialize};
+
+#[allow(clippy::large_enum_variant)]
+#[must_use]
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub enum QueryEvent {
+    None,
+    Debug {
+        message: String,
+    },
+
+    Done,
+
+    MakeExternal {
+        instance_id: u64,
+        constructor: Term,
+    },
+
+    ExternalCall {
+        /// Persistent id across all requests for results from the same external call.
+        call_id: u64,
+        /// The external instance to make this call on. None for functions or constructors.
+        instance: Option<Term>,
+        /// Field name to lookup or method name to call. A class name indicates a constructor
+        /// should be called.
+        attribute: Symbol,
+        /// List of arguments to use if this is a method call.
+        args: Option<Vec<Term>>,
+    },
+
+    /// Checks if the instance is an instance of (a subclass of) the class_tag.
+    ExternalIsa {
+        call_id: u64,
+        instance: Term,
+        class_tag: Symbol,
+    },
+
+    /// Checks if the instance is more specifically and instance/subclass of A than B.
+    ExternalIsSubSpecializer {
+        call_id: u64,
+        instance_id: u64,
+        left_class_tag: Symbol,
+        right_class_tag: Symbol,
+    },
+
+    /// Unifies two external instances.
+    ExternalUnify {
+        call_id: u64,
+        left_instance_id: u64,
+        right_instance_id: u64,
+    },
+
+    Result {
+        bindings: Bindings,
+        trace: Option<TraceResult>,
+    },
+
+    ExternalOp {
+        call_id: u64,
+        operator: Operator,
+        args: TermList,
+    },
+}
