+++ polar-core/src/macros.rs
@@ -8,8 +8,8 @@
 use std::collections::BTreeMap;
 use std::sync::atomic::{AtomicU64, Ordering};
 
+use crate::rules::*;
 use crate::terms::*;
-use crate::types::*;
 
 pub const ORD: Ordering = Ordering::SeqCst;
 pub static NEXT_ID: AtomicU64 = AtomicU64::new(0);
