+++ polar-core/src/parser.rs
@@ -8,8 +8,8 @@ lalrpop_mod!(
 
 use super::error::{self, PolarResult};
 use super::lexer::{self, Lexer};
+use super::rules::*;
 use super::terms::*;
-use super::types::*;
 
 #[derive(Clone, Debug, PartialEq)]
 pub enum Line {
