+++ polar-core/src/polar.lalrpop
@@ -8,7 +8,8 @@ use std::collections::{HashMap, BTreeMap};
 use crate::lexer::{self, Token};
 use crate::parser::Line;
 use crate::error;
-use crate::types::*;
+use crate::terms::*;
+use crate::rules::*;
 use crate::terms::*;
 use crate::numerics::*;
