+++ polar-core/src/rewrites.rs
@@ -1,5 +1,6 @@
+use super::kb::*;
+use super::rules::*;
 use super::terms::*;
-use super::types::*;
 
 /// Replace the left value by the AND of the right and the left.
 fn and_wrap(a: &mut Term, b: Term) {
@@ -9,6 +10,16 @@ fn and_wrap(a: &mut Term, b: Term) {
     });
 
     a.replace_value(new_value);
+}
+
+pub fn unwrap_and(term: Term) -> TermList {
+    match term.value() {
+        Value::Expression(Operation {
+            operator: Operator::And,
+            args,
+        }) => args.clone(),
+        _ => vec![term.clone()],
+    }
 }
 
 /// Checks if the expression needs to be rewritten. If so,
