+++ polar-core/src/rules.rs
@@ -1,0 +1,162 @@
+use super::terms::*;
+use serde::{Deserialize, Serialize};
+use std::collections::{BTreeMap, BTreeSet, HashMap};
+use std::sync::atomic::{AtomicU64, Ordering};
+use std::sync::Arc;
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct Parameter {
+    pub parameter: Term,
+    pub specializer: Option<Term>,
+}
+
+impl Parameter {
+    pub fn map_replace<F>(&mut self, f: &mut F)
+    where
+        F: FnMut(&Term) -> Term,
+    {
+        self.parameter.map_replace(f);
+        self.specializer.iter_mut().for_each(|p| p.map_replace(f));
+    }
+
+    pub fn is_ground(&self) -> bool {
+        self.specializer.is_none() && self.parameter.value().is_ground()
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct Rule {
+    pub name: Symbol,
+    pub params: Vec<Parameter>,
+    pub body: Term,
+}
+
+impl Rule {
+    pub fn map_replace<F>(&mut self, f: &mut F)
+    where
+        F: FnMut(&Term) -> Term,
+    {
+        self.params.iter_mut().for_each(|p| p.map_replace(f));
+        self.body.map_replace(f);
+    }
+
+    pub fn is_ground(&self) -> bool {
+        self.params.iter().all(|p| p.is_ground())
+    }
+}
+
+pub type Rules = Vec<Arc<Rule>>;
+
+type RuleSet = BTreeSet<u64>;
+
+#[derive(Clone, Default, Debug)]
+struct RuleIndex {
+    rules: RuleSet,
+    index: HashMap<Option<Value>, RuleIndex>,
+}
+
+impl RuleIndex {
+    pub fn index_rule(&mut self, rule_id: u64, params: &[Parameter], i: usize) {
+        if i < params.len() {
+            self.index
+                .entry({
+                    if params[i].is_ground() {
+                        Some(params[i].parameter.value().clone())
+                    } else {
+                        None
+                    }
+                })
+                .or_insert_with(RuleIndex::default)
+                .index_rule(rule_id, params, i + 1);
+        } else {
+            self.rules.insert(rule_id);
+        }
+    }
+
+    #[allow(clippy::comparison_chain)]
+    pub fn get_applicable_rules(&self, args: &[Term], i: usize) -> RuleSet {
+        if i < args.len() {
+            // Check this argument and recurse on the rest.
+            let filter_next_args =
+                |index: &RuleIndex| -> RuleSet { index.get_applicable_rules(args, i + 1) };
+            let arg = args[i].value();
+            if arg.is_ground() {
+                // Check the index for a ground argument.
+                let mut ruleset = self
+                    .index
+                    .get(&Some(arg.clone()))
+                    .map(|index| filter_next_args(index))
+                    .unwrap_or_else(RuleSet::default);
+
+                // Extend for a variable parameter.
+                if let Some(index) = self.index.get(&None) {
+                    ruleset.extend(filter_next_args(index));
+                }
+                ruleset
+            } else {
+                // Accumulate all indexed arguments.
+                self.index.values().fold(
+                    RuleSet::default(),
+                    |mut result: RuleSet, index: &RuleIndex| {
+                        result.extend(filter_next_args(index).into_iter());
+                        result
+                    },
+                )
+            }
+        } else {
+            // No more arguments.
+            self.rules.clone()
+        }
+    }
+}
+
+#[derive(Clone)]
+pub struct GenericRule {
+    pub name: Symbol,
+    rules: HashMap<u64, Arc<Rule>>,
+    index: RuleIndex,
+    next_rule_id: u64,
+}
+
+impl GenericRule {
+    pub fn new(name: Symbol, rules: Rules) -> Self {
+        let mut generic_rule = Self {
+            name,
+            rules: Default::default(),
+            index: Default::default(),
+            next_rule_id: 0,
+        };
+
+        for rule in rules {
+            generic_rule.add_rule(rule);
+        }
+
+        generic_rule
+    }
+
+    pub fn add_rule(&mut self, rule: Arc<Rule>) {
+        let rule_id = self.next_rule_id();
+
+        assert!(
+            self.rules.insert(rule_id, rule.clone()).is_none(),
+            "Rule id already used."
+        );
+        self.index.index_rule(rule_id, &rule.params[..], 0);
+    }
+
+    #[allow(clippy::ptr_arg)]
+    pub fn get_applicable_rules(&self, args: &TermList) -> Rules {
+        self.index
+            .get_applicable_rules(&args, 0)
+            .iter()
+            .map(|id| self.rules.get(id).expect("Rule missing"))
+            .cloned()
+            .collect()
+    }
+
+    fn next_rule_id(&mut self) -> u64 {
+        let v = self.next_rule_id;
+        self.next_rule_id += 1;
+        v
+    }
+}
