+++ polar-core/src/traces.rs
@@ -1,0 +1,23 @@
+use super::rules::*;
+use super::terms::*;
+use serde::{Deserialize, Serialize};
+use std::rc::Rc;
+use std::sync::Arc;
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+pub enum Node {
+    Rule(Arc<Rule>),
+    Term(Term),
+}
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+pub struct Trace {
+    pub node: Node,
+    pub children: Vec<Rc<Trace>>,
+}
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+pub struct TraceResult {
+    pub trace: Rc<Trace>,
+    pub formatted: String,
+}
