+++ polar-core/src/warnings.rs
@@ -1,6 +1,7 @@
 use super::formatting::source_lines;
+use super::kb::*;
+use super::rules::*;
 use super::terms::*;
-use super::types::*;
 
 use std::collections::{hash_map::Entry, HashMap};
